infrastructure/cdk/requirements.txt
aws-cdk-lib>=2.170.0
constructs>=10.3.0
boto3>=1.34.0
python-dotenv>=1.0.0
